remove aptos-move/e2e-tests-replay
Removing this per offline discussion.
Consolidate change set conversion logic into SessionOut.

Closes: #846
