--- aptos-move/e2e-tests-replay/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "aptos-e2e-tests-replay"
-version = "0.1.0"
-authors = ["Aptos Labs <opensource@aptoslabs.com>"]
-description = "A tool that reconstructs and replays test cases from the trace dump of E2E tests"
-license = "Apache-2.0"
-edition = "2018"
-publish = false
-
-[dependencies]
-anyhow = "1.0.52"
-bcs = "0.1.2"
-structopt = "0.3.21"
-walkdir = "2.3.1"
-
-aptos-workspace-hack = { path = "../../crates/aptos-workspace-hack" }
-move-deps = { path = "../move-deps", features=["address32"] }
-aptos-types = { path = "../../types", features = ["fuzzing"] }
-framework =  { path = "../framework" }
-aptos-vm = { path = "../../aptos-move/aptos-vm" }
-language-e2e-tests = { path = "../e2e-tests" }
