+++ x.toml
@@ -167,7 +167,6 @@ members = [
     # Please keep this list in alphabetical order!
 
     "testcases",
-    "aptos-e2e-tests-replay",
     "aptos-fuzz",
     "aptos-fuzzer",
     "aptos-keygen",
